+++ sparse_strips/vello_cpu/tests/basic.rs
@@ -70,7 +70,7 @@ fn empty_1134x1376() {
 fn full_cover_1() {
     let mut ctx = get_ctx(8, 8, true);
 
-    ctx.set_paint(BEIGE.into());
+    ctx.set_paint(BEIGE);
     ctx.fill_path(&Rect::new(0.0, 0.0, 8.0, 8.0).to_path(0.1));
 
     check_ref(&ctx, "full_cover_1");
@@ -90,7 +90,7 @@ fn filled_triangle() {
         path
     };
 
-    ctx.set_paint(LIME.into());
+    ctx.set_paint(LIME);
     ctx.fill_path(&path);
 
     check_ref(&ctx, "filled_triangle");
@@ -110,7 +110,7 @@ fn stroked_triangle() {
     };
 
     ctx.set_stroke(Stroke::new(3.0));
-    ctx.set_paint(LIME.into());
+    ctx.set_paint(LIME);
     ctx.stroke_path(&path);
 
     check_ref(&ctx, "stroked_triangle");
@@ -120,7 +120,7 @@ fn stroked_triangle() {
 fn filled_circle() {
     let mut ctx = get_ctx(100, 100, false);
     let circle = Circle::new((50.0, 50.0), 45.0);
-    ctx.set_paint(LIME.into());
+    ctx.set_paint(LIME);
     ctx.fill_path(&circle.to_path(0.1));
 
     check_ref(&ctx, "filled_circle");
@@ -131,7 +131,7 @@ fn filled_overflowing_circle() {
     let mut ctx = get_ctx(100, 100, false);
     let circle = Circle::new((50.0, 50.0), 50.0 + 1.0);
 
-    ctx.set_paint(LIME.into());
+    ctx.set_paint(LIME);
     ctx.fill_path(&circle.to_path(0.1));
 
     check_ref(&ctx, "filled_overflowing_circle");
@@ -142,7 +142,7 @@ fn filled_fully_overflowing_circle() {
     let mut ctx = get_ctx(100, 100, false);
     let circle = Circle::new((50.0, 50.0), 80.0);
 
-    ctx.set_paint(LIME.into());
+    ctx.set_paint(LIME);
     ctx.fill_path(&circle.to_path(0.1));
 
     check_ref(&ctx, "filled_fully_overflowing_circle");
@@ -153,7 +153,7 @@ fn filled_circle_with_opacity() {
     let mut ctx = get_ctx(100, 100, false);
     let circle = Circle::new((50.0, 50.0), 45.0);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_path(&circle.to_path(0.1));
 
     check_ref(&ctx, "filled_circle_with_opacity");
@@ -165,7 +165,7 @@ fn filled_overlapping_circles() {
 
     for e in [(35.0, 35.0, RED), (65.0, 35.0, GREEN), (50.0, 65.0, BLUE)] {
         let circle = Circle::new((e.0, e.1), 30.0);
-        ctx.set_paint(e.2.with_alpha(0.5).into());
+        ctx.set_paint(e.2.with_alpha(0.5));
         ctx.fill_path(&circle.to_path(0.1));
     }
 
@@ -178,7 +178,7 @@ fn stroked_circle() {
     let circle = Circle::new((50.0, 50.0), 45.0);
     let stroke = Stroke::new(3.0);
 
-    ctx.set_paint(LIME.into());
+    ctx.set_paint(LIME);
     ctx.set_stroke(stroke);
     ctx.stroke_path(&circle.to_path(0.1));
 
@@ -191,7 +191,7 @@ fn rectangle_above_viewport() {
     let mut ctx = get_ctx(10, 10, false);
     let rect = Rect::new(2.0, -5.0, 8.0, 8.0);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "rectangle_above_viewport");
@@ -212,7 +212,7 @@ fn triangle_above_and_wider_than_viewport() {
         path
     };
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_path(&path);
 
     check_ref(&ctx, "triangle_above_and_wider_than_viewport");
@@ -225,7 +225,7 @@ fn rectangle_left_of_viewport() {
     let mut ctx = get_ctx(10, 10, false);
     let rect = Rect::new(-4.0, 3.0, 1.0, 8.0);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "rectangle_left_of_viewport");
@@ -236,7 +236,7 @@ fn filling_nonzero_rule() {
     let mut ctx = get_ctx(100, 100, false);
     let star = crossed_line_star();
 
-    ctx.set_paint(MAROON.into());
+    ctx.set_paint(MAROON);
     ctx.fill_path(&star);
 
     check_ref(&ctx, "filling_nonzero_rule");
@@ -247,7 +247,7 @@ fn filling_evenodd_rule() {
     let mut ctx = get_ctx(100, 100, false);
     let star = crossed_line_star();
 
-    ctx.set_paint(MAROON.into());
+    ctx.set_paint(MAROON);
     ctx.set_fill_rule(Fill::EvenOdd);
     ctx.fill_path(&star);
 
@@ -259,7 +259,7 @@ fn filled_aligned_rect() {
     let mut ctx = get_ctx(30, 20, false);
     let rect = Rect::new(1.0, 1.0, 29.0, 19.0);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_aligned_rect");
@@ -275,7 +275,7 @@ fn stroked_unaligned_rect() {
         ..Default::default()
     };
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -292,7 +292,7 @@ fn stroked_unaligned_rect_as_path() {
         ..Default::default()
     };
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_path(&rect);
 
@@ -305,7 +305,7 @@ fn stroked_aligned_rect() {
     let rect = Rect::new(5.0, 5.0, 25.0, 25.0);
     let stroke = miter_stroke_2();
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -322,7 +322,7 @@ fn overflowing_stroked_rect() {
         ..Default::default()
     };
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -335,7 +335,7 @@ fn round_stroked_rect() {
     let rect = Rect::new(5.0, 5.0, 25.0, 25.0);
     let stroke = Stroke::new(3.0);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -352,7 +352,7 @@ fn bevel_stroked_rect() {
         ..Default::default()
     };
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -364,7 +364,7 @@ fn filled_unaligned_rect() {
     let mut ctx = get_ctx(30, 20, false);
     let rect = Rect::new(1.5, 1.5, 28.5, 18.5);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_unaligned_rect");
@@ -376,7 +376,7 @@ fn filled_transformed_rect_1() {
     let rect = Rect::new(0.0, 0.0, 10.0, 10.0);
 
     ctx.set_transform(Affine::translate((10.0, 10.0)));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_transformed_rect_1");
@@ -388,7 +388,7 @@ fn filled_transformed_rect_2() {
     let rect = Rect::new(5.0, 5.0, 10.0, 10.0);
 
     ctx.set_transform(Affine::scale(2.0));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_transformed_rect_2");
@@ -400,7 +400,7 @@ fn filled_transformed_rect_3() {
     let rect = Rect::new(0.0, 0.0, 10.0, 10.0);
 
     ctx.set_transform(Affine::new([2.0, 0.0, 0.0, 2.0, 5.0, 5.0]));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_transformed_rect_3");
@@ -415,7 +415,7 @@ fn filled_transformed_rect_4() {
         45.0 * PI / 180.0,
         Point::new(15.0, 15.0),
     ));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_transformed_rect_4");
@@ -428,7 +428,7 @@ fn stroked_transformed_rect_1() {
     let stroke = miter_stroke_2();
 
     ctx.set_transform(Affine::translate((10.0, 10.0)));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -442,7 +442,7 @@ fn stroked_transformed_rect_2() {
     let stroke = miter_stroke_2();
 
     ctx.set_transform(Affine::scale(2.0));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -456,7 +456,7 @@ fn stroked_transformed_rect_3() {
     let stroke = miter_stroke_2();
 
     ctx.set_transform(Affine::new([2.0, 0.0, 0.0, 2.0, 5.0, 5.0]));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -473,7 +473,7 @@ fn stroked_transformed_rect_4() {
         45.0 * PI / 180.0,
         Point::new(15.0, 15.0),
     ));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&rect);
 
@@ -485,7 +485,7 @@ fn strip_inscribed_rect() {
     let mut ctx = get_ctx(30, 20, false);
     let rect = Rect::new(1.5, 9.5, 28.5, 11.5);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "strip_inscribed_rect");
@@ -496,7 +496,7 @@ fn filled_vertical_hairline_rect() {
     let mut ctx = get_ctx(5, 8, false);
     let rect = Rect::new(2.25, 0.0, 2.75, 8.0);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_vertical_hairline_rect");
@@ -507,7 +507,7 @@ fn filled_vertical_hairline_rect_2() {
     let mut ctx = get_ctx(10, 10, false);
     let rect = Rect::new(4.5, 0.5, 5.5, 9.5);
 
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.fill_rect(&rect);
 
     check_ref(&ctx, "filled_vertical_hairline_rect_2");
@@ -520,7 +520,7 @@ fn filled_glyphs() {
     let (font, glyphs) = layout_glyphs("Hello, world!", font_size);
 
     ctx.set_transform(Affine::translate((0., f64::from(font_size))));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.glyph_run(&font)
         .font_size(font_size)
         .fill_glyphs(glyphs.into_iter());
@@ -535,7 +535,7 @@ fn stroked_glyphs() {
     let (font, glyphs) = layout_glyphs("Hello, world!", font_size);
 
     ctx.set_transform(Affine::translate((0., f64::from(font_size))));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.glyph_run(&font)
         .font_size(font_size)
         .stroke_glyphs(glyphs.into_iter());
@@ -550,7 +550,7 @@ fn skewed_glyphs() {
     let (font, glyphs) = layout_glyphs("Hello, world!", font_size);
 
     ctx.set_transform(Affine::translate((0., f64::from(font_size))));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.glyph_run(&font)
         .font_size(font_size)
         .glyph_transform(Affine::skew(-20_f64.to_radians().tan(), 0.0))
@@ -566,7 +566,7 @@ fn scaled_glyphs() {
     let (font, glyphs) = layout_glyphs("Hello,\nworld!", font_size);
 
     ctx.set_transform(Affine::translate((0., f64::from(font_size))).then_scale(2.0));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
+    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
     ctx.glyph_run(&font)
         .font_size(font_size)
         .fill_glyphs(glyphs.into_iter());
@@ -636,11 +636,11 @@ fn oversized_star() {
     // Center it in the middle of the viewport
     let star_path = circular_star(Point::new(50., 50.), 10, 30., 90.);
 
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_path(&star_path);
 
     let stroke = Stroke::new(2.0);
-    ctx.set_paint(DARK_BLUE.into());
+    ctx.set_paint(DARK_BLUE);
     ctx.set_stroke(stroke);
     ctx.stroke_path(&star_path);
 
@@ -658,14 +658,14 @@ fn clip_triangle_with_star() {
     triangle_path.close_path();
 
     let stroke = Stroke::new(1.0);
-    ctx.set_paint(DARK_BLUE.into());
+    ctx.set_paint(DARK_BLUE);
     ctx.set_stroke(stroke);
     ctx.stroke_path(&triangle_path);
 
     let star_path = circular_star(Point::new(50., 50.), 13, 25., 45.);
 
     ctx.clip(&star_path);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_path(&triangle_path);
     ctx.finish();
 
@@ -685,7 +685,7 @@ fn clip_rectangle_with_star_nonzero() {
     ctx.clip(&star_path);
     // Draw a rectangle that should be clipped by the star
     // The NonZero fill rule will treat self-intersecting regions as filled
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
     check_ref(&ctx, "clip_rectangle_with_star_nonzero");
@@ -704,7 +704,7 @@ fn clip_rectangle_with_star_evenodd() {
     ctx.clip(&star_path);
     // Draw a rectangle that should be clipped by the star
     // The EvenOdd rule should create a "hole" in the middle where the paths overlap
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
 
@@ -725,7 +725,7 @@ fn clip_rectangle_and_circle() {
 
     // Draw outlines of our clipping regions to visualize them
     let stroke = Stroke::new(1.0);
-    ctx.set_paint(DARK_BLUE.into());
+    ctx.set_paint(DARK_BLUE);
     ctx.set_stroke(stroke);
     ctx.stroke_rect(&clip_rect);
     ctx.stroke_path(&clip_circle);
@@ -736,7 +736,7 @@ fn clip_rectangle_and_circle() {
 
     // Then a filled rectangle that covers most of the canvas
     let large_rect = Rect::new(0.0, 0.0, 100.0, 100.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&large_rect);
     ctx.finish();
     check_ref(&ctx, "clip_rectangle_and_circle");
@@ -756,7 +756,7 @@ fn clip_with_translation() {
 
     // Draw a rectangle that should be clipped
     let rect = Rect::new(0.0, 0.0, 100.0, 100.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
     check_ref(&ctx, "clip_with_translation");
@@ -775,7 +775,7 @@ fn clip_with_scale() {
 
     // Draw a rectangle that should be clipped
     let rect = Rect::new(0.0, 0.0, 100.0, 100.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
 
@@ -798,7 +798,7 @@ fn clip_with_rotate() {
 
     // Draw a rectangle that should be clipped
     let rect = Rect::new(20.0, 20.0, 80.0, 80.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
 
@@ -822,7 +822,7 @@ fn clip_transformed_rect() {
 
     // Draw a smaller rectangle that should be clipped
     let rect = Rect::new(20.0, 20.0, 80.0, 80.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
 
@@ -854,7 +854,7 @@ fn clip_with_multiple_transforms() {
 
     // Draw a rectangle that should be clipped by both regions
     let rect = Rect::new(0.0, 0.0, 100.0, 100.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
     ctx.finish();
 
@@ -882,7 +882,7 @@ fn clip_with_save_restore() {
 
     // Draw a rectangle that should be clipped by both regions
     let rect = Rect::new(0.0, 0.0, 100.0, 100.0);
-    ctx.set_paint(REBECCA_PURPLE.into());
+    ctx.set_paint(REBECCA_PURPLE);
     ctx.fill_rect(&rect);
 
     // Restore to state before second clip
@@ -890,7 +890,7 @@ fn clip_with_save_restore() {
 
     // Draw another rectangle that should only be clipped by the first region
     let rect2 = Rect::new(0.0, 0.0, 100.0, 100.0);
-    ctx.set_paint(DARK_GREEN.with_alpha(0.5).into());
+    ctx.set_paint(DARK_GREEN.with_alpha(0.5));
     ctx.fill_rect(&rect2);
     ctx.finish();
     check_ref(&ctx, "clip_with_save_restore");
@@ -898,7 +898,7 @@ fn clip_with_save_restore() {
 
 fn draw_clipping_outline(ctx: &mut RenderContext, path: &BezPath) {
     let stroke = Stroke::new(1.0);
-    ctx.set_paint(DARK_BLUE.into());
+    ctx.set_paint(DARK_BLUE);
     ctx.set_stroke(stroke);
     ctx.stroke_path(path);
 }
@@ -922,7 +922,7 @@ fn bevel_stroke_2() -> Stroke {
 fn compose_destination() -> RenderContext {
     let mut ctx = get_ctx(50, 50, true);
     let rect = Rect::new(4.5, 4.5, 35.5, 35.5);
-    ctx.set_paint(YELLOW.with_alpha(0.35).into());
+    ctx.set_paint(YELLOW.with_alpha(0.35));
     ctx.set_stroke(bevel_stroke_2());
     ctx.fill_rect(&rect);
 
@@ -931,7 +931,7 @@ fn compose_destination() -> RenderContext {
 
 fn compose_source(ctx: &mut RenderContext) {
     let rect = Rect::new(14.5, 14.5, 45.5, 45.5);
-    ctx.set_paint(DARK_GREEN.with_alpha(0.8).into());
+    ctx.set_paint(DARK_GREEN.with_alpha(0.8));
     ctx.fill_rect(&rect);
 }
